deleted unnecessary stuff and modified filenames/imports to be more consistent
