+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import Header from './components/Header'
-import Main from './components/Main'
+import Header from './components/Header/Header.js'
+import Main from './components/Main.js'
 import './App.css'
 import {Route} from "react-router-dom"
