+++ client/src/components/Cupboard/Cupboard.css
@@ -1,0 +1,5 @@
+.Cupboard {
+    border: 1px solid black;
+    box-shadow: 1px 1px 1px 0px rgba(0,0,0,0.25);
+    border-radius: 65px;
+  }
