--- client/src/components/Header.js
@@ -1,15 +1,0 @@
-import React from "react";
-// import { Link } from 'react-router-dom'
-import NewItemButton from './NewItem/NewItem';
-import logo from "../logo.svg";
-
-const Header = () => (
-  <div>
-    <header>
-      <img src={logo} className="App-logo" alt="logo" />
-    </header>
-
-  </div>
-);
-
-export default Header;
