+++ client/src/components/Header/Header.js
@@ -1,0 +1,10 @@
+import React from "react";
+// import { Link } from 'react-router-dom'
+
+const Header = () => (
+  <div>
+    
+  </div>
+);
+
+export default Header;
