+++ client/src/components/Main.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Switch, Route, Redirect } from "react-router-dom";
-import Home from "./Home/Home";
+import Home from "./Home/Home.js";
 
 // The Main component renders routes
