+++ client/src/components/Wrapper/Wrapper.css
@@ -1,6 +1,5 @@
 .wrapper {
     padding-top: 50px;
-    background: lightgrey;
     height: 100%;
     padding: 15px;
     justify-content: space-around;
