+++ client/src/components/Wrapper/Wrapper.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./style.css"
+import "./Wrapper.css"
 
 function Wrapper(props) {
   return <div className="wrapper">{props.children}</div>;
